Moved validators and updated default
